remove useless metrics file
