--- cmd/security-agent/app/metrics.go
@@ -1,31 +1,0 @@
-// Unless explicitly stated otherwise all files in this repository are licensed
-// under the Apache License Version 2.0.
-// This product includes software developed at Datadog (https://www.datadoghq.com/).
-// Copyright 2016-present Datadog, Inc.
-
-package app
-
-import (
-	"fmt"
-	"time"
-
-	ddgostatsd "github.com/DataDog/datadog-go/v5/statsd"
-
-	"github.com/DataDog/datadog-agent/pkg/version"
-)
-
-// sendRunningMetrics exports a metric to distinguish between security-agent modules that are activated
-func sendRunningMetrics(statsdClient *ddgostatsd.Client, moduleName string) *time.Ticker {
-	// Retrieve the agent version using a dedicated package
-	tags := []string{fmt.Sprintf("version:%s", version.AgentVersion)}
-
-	// Send the metric regularly
-	heartbeat := time.NewTicker(15 * time.Second)
-	go func() {
-		for range heartbeat.C {
-			statsdClient.Gauge(fmt.Sprintf("datadog.security_agent.%s.running", moduleName), 1, tags, 1) //nolint:errcheck
-		}
-	}()
-
-	return heartbeat
-}
